Config for custom currency
